get issues based on project id and user id. Tests improved

+++ src/controllers/issue.ts
@@ -3,12 +3,25 @@ import StatusCodes from 'http-status-codes';
 import Issue from '../models/issue';
 
 export const issues = async (req: Request, res: Response, next: Next): Promise<void> => {
-  const issues = await Issue.find({});
+  const projectId = req.query.projectId;
+
+  if (!projectId) {
+    res.sendStatus(StatusCodes.BAD_REQUEST);
+    return
+  }
+
+  const issues = await Issue.find({
+    projectId: projectId,
+    createdBy: req.user['_id'].toString()
+  });
   res.send(issues);
 };
 
 export const issue = async (req: Request, res: Response, next: Next): Promise<void> => {
-  const issue: any = await Issue.findOne({ _id: req.params.id });
+  const issue: any = await Issue.findOne({
+    _id: req.params.id,
+    createdBy: req.user['_id'].toString()
+  });
   res.send(issue);
 };
 
@@ -19,7 +32,16 @@ export const save = async (req: Request, res: Response, next: Next): Promise<voi
 };
 
 export const update = async (req: Request, res: Response, next: Next): Promise<void> => {
-  const issue = await Issue.findOne({ _id: req.params.id });
+  const issue = await Issue.findOne({
+    _id: req.params.id,
+    createdBy: req.user['_id'].toString()
+  });
+
+  if (!issue) {
+    res.sendStatus(StatusCodes.NOT_FOUND);
+    return
+  }
+
   issue.title = req.body.title;
   issue.content = req.body.content;
   const savedIssue = await issue.save();
@@ -27,6 +49,6 @@ export const update = async (req: Request, res: Response, next: Next): Promise<v
 };
 
 export const remove = async (req: Request, res: Response, next: Next): Promise<void> => {
-  await Issue.deleteOne({ _id: req.params.id });
+  await Issue.deleteOne({ _id: req.params.id, createdBy: req.user['_id'].toString() });
   res.sendStatus(StatusCodes.OK);
 }
